qcacmn: Return proper error status from modules under hif_start
Some of the HIF API's that do buffer allocations and operations
return either 0 or 1 as return status. Information is lost since
the same value is returned in case of buffer allocation failures,
queue operation failure or anything else. Caller can't decide if
to perform recovery, graceful failure or BUG_ON.
Return the appropriate error status to the callers.

Change-Id: Idccd3968231d030311fa5581aed78849b729227d
CRs-Fixed: 2105913
Bug: 65405896
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>