usb: dwc3-msm: Switch to freezable workqueue for sm_work
Use freezable workqueue for running the sm_work. The sm_work
add/removes the platform_dev to switch between host and the
gadget mode. However, the work_item does not have protection
against the system suspend/resume. Using freezable workqueue
ensures that the device tree is not changed while the
system is suspending or resuming. Using the system freezable
workqueue instead of creating a new one.

Bug: 65638337
Change-Id: I9242e9386043cea4f8f765010e32efaa3f10059d
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
[@0ctobot: Adapted for kernel.lnx.4.4.r35-rel]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>