[SQUASH] Revert "simple_lmk: Introduce Simple LMK (Low Memory Killer) for Android"
This appears to be contributing to some rogue task killing resulting in lockups and freezes as outlined in 0ctobot/neutrino_kernel_essential_msm8998#1. Whether SLMK is directly responsible for this, or some other deficiency in source is exacerbating the issue following it's inclusion any benefits attained from switching are currently null and void.

This reverts the following commits:
  b9f4926 ("FIXUP: simple_lmk: Introduce Simple LMK (Low Memory Killer) for Android")
  0c11eac ("simple_lmk: Boost CPU and DDR bus when mem reclaim is invoked")
  93c5f67 ("simple_lmk: Introduce simple low memory killer for Android")

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>