Revert "USB: dwc3: hold wakelock for audio devices"
This was meant to hold the device awake while Headphone Audio Adapter is in use. However, this wakelock can potentially be triggered by various other USB devices and subsequently fails to be released. I've considered ammending this commit with YaroST12/MATA_Kernel@cec1c80 ("dwc3-msm: be more precise with wakelock") to specifically target the stock Essential dongle. However, this is also not ideal as it will break functionality for some third party Audio Adapters. Because of this, I will be attempting an alternative approach. If this also proves to be ineffective I may revist this commit in the future.

This reverts commit b414aab914c3534032a6bc764fb0f71c69975651.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>