kernel: Boost CPU to the max when Quickstep becomes a top app
Android moves the Quickstep process to the top-app cpuset when the
multitasking view is entered since this launcher handles the multitasking
interface, which is very expensive. Because of this, when rapidly
hitting the multitask button and switching apps sometimes there's
noticeably bad stuttering due to the overhead from the app-switch
interface itself.

In order to fix the stuttering when multitasking, boost all CPUs and the
DRAM bus to the max when Quickstep is moved to the top-app cpuset.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
[@0ctobot: Adapted for mata from sultanxda/android_kernel_google_wahoo@d305c75 and sultanxda/android_kernel_google_wahoo@d7ef260]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>